one last function still not working

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE department (
 CREATE TABLE role (
     id INT AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR (30) UNIQUE NULL,
-    salary DECIMAL NOT NULL,
+    salary INT NOT NULL,
     department_id INT NOT NULL,
     INDEX dep_ind (department_id),
     CONSTRAINT fk_department FOREIGN KEY (department_id) REFERENCES department(id)
